refactor: Member 도메인 디렉토리 변경
