+++ server/EverybodyChachapark/src/main/java/com/DuTongChitongYutong/EverybodyChachapark/domain/member/controller/MemberController.java
@@ -1,5 +1,9 @@
-package com.DuTongChitongYutong.EverybodyChachapark.member;
+package com.DuTongChitongYutong.EverybodyChachapark.domain.member.controller;
 
+import com.DuTongChitongYutong.EverybodyChachapark.domain.member.dto.MemberDto;
+import com.DuTongChitongYutong.EverybodyChachapark.domain.member.mapper.MemberMapper;
+import com.DuTongChitongYutong.EverybodyChachapark.domain.member.service.MemberService;
+import com.DuTongChitongYutong.EverybodyChachapark.domain.member.entity.Member;
 import com.DuTongChitongYutong.EverybodyChachapark.response.SingleResponseDto;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
